binalize_spikes returns float64 so downstream matmuls hit BLAS gemm
X_t @ Theta2 in _fs_from_theta_dense was 10x slower with int64 storage
(0.42ms vs 0.04ms standalone) because numpy falls back to a non-BLAS
loop for int matmul. binalize_spikes now stores as float64; values are
still {0., 1.} so all source-line equations are literally unchanged.

Paired N=60 bench (15 reps):
  master: med=1.180  min=1.161  max=1.265
  opt:    med=1.126  min=1.095  max=1.144   (~-4.6% med, -5.7% min)

+++ transforms.py
@@ -238,8 +238,12 @@ def binalize_spikes(spikes, order, window):
     spikes = spikes[:int(T / window) * window,:,:]
     spikes = spikes.reshape((int(T / window), window, R, N))
     spikes = spikes.any(axis=1)
-    spikes = spikes.astype(int)
-    
+    # Store as float64 so downstream matmuls (X_t @ Theta2 in
+    # _fs_from_theta_dense, X_t.T @ res in pseudo_dllk, batched matmul
+    # in compute_y) dispatch to BLAS gemm. Values are still {0., 1.},
+    # so equation arithmetic is identical.
+    spikes = spikes.astype(numpy.float64)
+
     return spikes
 
 def compute_y(spikes, order):
